Refactor: Rename XSS scanner and improve stability
Rename the "Ultimate Modern XSS Scanner" to "Stable XSS Scanner" and enhance its reconnaissance and browser management for improved stability and reliability.
